[SME][RegAlloc] Reduce register weights of SME tiles
Offering: Bisheng

Because of the hierarchical order of register definition of SME tiles,
the weights that are calculated for the tile register are too huge.
This causes register pressure underflow error, because of the operations
performed on huge signed weights. This patches fixes the weights of
the tile registers by skipping the normalize method for weight
calculation.